Migrate the hybrid proxy setup to maven/connect
Now we use OSGi connect, there is no more need to split between
maven/OSGi world when doing proxy setup
