+++ p2-maven-plugin/pom.xml
@@ -175,6 +175,11 @@
 		    <artifactId>org.eclipse.equinox.simpleconfigurator.manipulator</artifactId>
 		    <version>2.2.0</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.platform</groupId>
+			<artifactId>org.eclipse.core.net</artifactId>
+			<version>1.3.1200</version>
+		</dependency>
 		<!-- ECF  -->
 		<dependency>
 			<groupId>org.eclipse.platform</groupId>
