+++ tycho-core/pom.xml
@@ -264,11 +264,6 @@
 			<version>3.25.0</version>
 		</dependency>
 		<dependency>
-			<groupId>org.eclipse.platform</groupId>
-			<artifactId>org.eclipse.core.net</artifactId>
-			<version>1.3.1200</version>
-		</dependency>
-		<dependency>
 			<groupId>biz.aQute.bnd</groupId>
 			<artifactId>biz.aQute.bndlib</artifactId>
 		</dependency>
